extract get all from model to repo, update logic and tests
