+++ controllers/games.js
@@ -1,8 +1,8 @@
-const Games = require("../models/games");
+const repo = require("../repos/games");
 
 const getAll = async (req, res) => {
   try {
-    res.json(await Games.getAll());
+    res.json(await repo.getAll());
   } catch (error) {
     throw error;
   }
