+++ tests/controllers/games.js
@@ -2,17 +2,17 @@ const chai = require("chai");
 const sinon = require("sinon");
 const expect = chai.expect;
 
-const model = require("../../models/games");
+const repo = require("../../repos/games");
 const controller = require("../../controllers/games");
 
 const { mockRequest, mockResponse } = require("./helpers");
 
 describe("GamesController", () => {
   describe("getAll", () => {
-    it("should call to model", async () => {
+    it("should call to repo", async () => {
       const req = mockRequest();
       const res = mockResponse();
-      const stub = sinon.stub(model, "getAll");
+      const stub = sinon.stub(repo, "getAll");
 
       await controller.getAll(req, res);
